refactor: separate server and client code for enabling translation in server components
