+++ src/index.ts
@@ -1,5 +1,6 @@
-export { TranslationProvider } from './TranslationProvider'
-export { useTranslations } from './useTranslations'
-export { createT } from './createT'
-export { getTranslations } from './getTranslations'
-export * from './types'
+export { getTranslations } from './server/getTranslations'
+
+export { TranslationProvider } from './client/TranslationProvider'
+export { useTranslations } from './client/useTranslations'
+
+export * from './shared/types'
